Moved type database to BindingsGeneration
